#272 coderabbit指摘対応
