+++ .gitignore
@@ -2,6 +2,7 @@
 /virtual-voicebot-backend/storage/recordings
 /virtual-voicebot-backend/storage/notifications/*.jsonl
 /virtual-voicebot-frontend/storage/recordings
+/virtual-voicebot-frontend/storage/notifications/notifications.json
 /virtual-voicebot-frontend/storage/db
 /virtual-voicebot-backend/advenc_calendar
 /virtual-voicebot-backend/spec
