+++ virtual-voicebot-backend/src/protocol/sip/mod.rs
@@ -14,6 +14,7 @@ pub mod transaction;
 pub mod transport;
 pub mod tx;
 pub mod types;
+pub mod utils;
 
 #[allow(unused_imports)]
 pub use message::{SipHeader, SipMessage, SipMethod, SipRequest, SipResponse};
@@ -28,3 +29,4 @@ pub use protocols::*;
 
 pub use core::{parse_offer_sdp, SipCore};
 pub use types::{SipCommand, SipConfig, SipEvent};
+pub use utils::extract_user_from_to;
